Reorganize Codebase
- [+] feat(config.go): add new config file to handle application configuration
- [+] refactor(run.go): remove Config struct and getConfig function, moved to config.go
